requirements.txt
amqp==5.1.1
    # via kombu
anyio==3.5.0
    # via starlette
asgiref==3.5.0
    # via uvicorn
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
celery==5.2.6
    # via
    #   -r requirements.in
    #   flower
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
deprecated==1.2.13
    # via redis
fastapi==0.75.2
    # via -r requirements.in
flower==1.0.0
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
humanize==4.0.0
    # via flower
idna==3.3
    # via
    #   anyio
    #   requests
kombu==5.2.4
    # via celery
lxml==4.8.0
    # via pytrends
numpy==1.22.3
    # via pandas
packaging==21.3
    # via redis
pandas==1.4.2
    # via
    #   -r requirements.in
    #   pytrends
prometheus-client==0.14.1
    # via flower
prompt-toolkit==3.0.29
    # via click-repl
pydantic==1.9.0
    # via fastapi
pyparsing==3.0.8
    # via packaging
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.20.0
    # via -r requirements.in
pytrends==4.8.0
    # via -r requirements.in
pytz==2022.1
    # via
    #   celery
    #   flower
    #   pandas
redis==4.2.2
    # via -r requirements.in
requests==2.27.1
    # via pytrends
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sniffio==1.2.0
    # via anyio
starlette==0.17.1
    # via fastapi
tornado==6.1
    # via flower
typing-extensions==4.2.0
    # via pydantic
urllib3==1.26.9
    # via requests
uvicorn==0.17.6
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wrapt==1.14.0
    # via deprecated
